[#98] S3 moving: small updates for downloading/uploading
