+++ Pipfile
@@ -18,10 +18,7 @@ alembic = "==1.7.5"
 PyJWT = "==2.3.0"
 httpx = "==0.21.1"
 redis = "==4.0.2"
-#TODO: move to aioboto3
-#boto3 = "==1.21.18"
-aioboto3 =  "==9.4.0"
-tqdm = "==4.63.0"
+boto3 = "==1.21.18"
 youtube_dl = "==2021.12.17"
 sentry-sdk = "==1.5.0"
 python-dotenv = "==0.19.2"
